Modif méthode sur les components
