+++ src/assets/styles/navbar/navbar.css
@@ -2,7 +2,7 @@ nav.navbar {
     background-color: #023859;
 }
 
-img{
+.img-logo{
     height: 7rem;
 }
