+++ src/components/LoginComponents.js
@@ -1,8 +1,45 @@
+import AccountService from "../services/AccountService.js";
+import App from "../scripts/App.js";
+import LoginService from "../services/LoginService.js";
+
 export default class LoginComponent extends HTMLElement {
   constructor() {
     super();
     this.innerHTML = this.render();
+    this.querySelector("form").onsubmit = this.handleLoginFormSubmit;
   }
+
+  handleLoginFormSubmit = (e) => {
+    e.preventDefault();
+    const entries = Object.fromEntries(new FormData(e.target));
+    const accountService = new AccountService();
+    const accounts = accountService.read(
+      (account) => account.email === entries.email
+    );
+
+    if (accounts.length === 0) {
+      const noExistingUser = document.getElementById("controlLoginAccount");
+      noExistingUser.textContent = "Compte inexistant ou mdp incorrect !";
+      setTimeout(function () {
+        noExistingUser.textContent = "";
+      }, 2000);
+      return;
+    }
+    const currentUser = accounts.pop();
+    if (currentUser.password === entries.loginPassword) {
+      LoginService.login(currentUser);
+      alert("Connecté !");
+      history.pushState(null, null, "/");
+      App.instance.router.navigate();
+      return;
+    } else {
+      const wrongPassword = document.getElementById("controlLoginAccount");
+      wrongPassword.textContent = "Compte inexistant ou mdp incorrect !";
+      setTimeout(function () {
+        wrongPassword.textContent = "";
+      }, 2000);
+    }
+  };
 
   render() {
     return `<style>@import "./src/assets/styles/loginPageStyle/login.css"</style>
